requirements.txt
gssapi
djangorestframework
